Add integration tests for the person detection rune

+++ runic-types/src/wasm32/capability.rs
@@ -4,7 +4,7 @@ use alloc::vec::Vec;
 
 pub type Random<T> = GenericCapability<T, { crate::capabilities::RAND }>;
 pub type Accelerometer = GenericCapability<f32, { crate::capabilities::ACCEL }>;
-pub type Image = GenericCapability<u32, { crate::capabilities::IMAGE }>;
+pub type Image = GenericCapability<u8, { crate::capabilities::IMAGE }>;
 pub type Sound = GenericCapability<i16, { crate::capabilities::SOUND }>;
 pub type Raw<T> = GenericCapability<T, { crate::capabilities::RAW }>;
